requirements.txt
aiofiles
aioredis==1.3.1
aiosqlite==0.11.0                        # indirect dependency of tortoise-orm
backtrader==1.9.74.123
click==7.0
colorama
fastapi==0.45.0
fastapi-users[tortoise-orm]
httpx==0.10.1
httpx-oauth
jinja2
python-multipart
pytest-asyncio==0.10.0
python-rapidjson==0.9.1
python-dotenv==0.10.5
pandas==0.25.3
redis
requests==2.22.0
simplejson==3.17.0
stackprinter==0.2.3
structlog==20.1.0
tortoise-orm==0.15.23
typesystem==0.2.4
ujson==1.35
uvicorn==0.11.1
websockets==8.1                          # direct dependecy + required by uvicorn
snakeviz
# TA-Lib